pyproject.toml
[tool.poetry]
name = "helios"
version = "0.1.0"
description = ""
authors = ["MMartin09 <mmartin09@outlook.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
fastapi = "^0.104.0"
pydantic-settings = "^2.0.3"
loguru = "^0.7.2"
tortoise-orm = {extras = ["asyncmy"], version = "^0.20.0"}
httpx = "^0.25.0"
influxdb-client = "^1.38.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Based on the config of FastAPI
[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]
